requirements.txt
haystack-ai
ollama-haystack
sentence-transformers